Update
- Dodano nowe role (moderator, zbanowany)
- Moderatorzy nie mogą edytować kont administratorów i innych moderatorów ani ustawiać takich rang innym użytkownikom. Mogą jedynie edytować dane kont użytkowników i zbanowanych (i zmieniać im range z jednej na drugą)
- Zbanowani nie będą mieli możliwości dodawania nowych miejsc (do zaimplementowania wraz z rozwojem projektu)
- Dodano obsługę glyphiconów z bootstrapa
- Podzielono menu na nowe role. Moderatorzy nie mają już wglądu do logów, debuggera i xmla, nie mogą też usuwać użytkowników

+++ app/controllers/UserManagerControl.class.php
@@ -130,6 +130,10 @@ class UserManagerControl
                 App::getSmarty()->assign("userDetails", $this->user);
                 break;
             case "delete" :
+                if(SessionUtils::loadData('role', true) == 'moderator'){
+                    Utils::addErrorMessage("Tylko administrator może usuwać użytkowników!");
+                    break;
+                }
                 $this->deleteUser($user_id);
                 break;
             case "edit" :
